mocks updating

+++ src/Ativos.Domain/Repositories/Usuarios/IUsuariosReadOnlyRepository.cs
@@ -4,7 +4,7 @@ public interface IUsuariosReadOnlyRepository
 {
     Task<List<Entities.Usuario>> GetAll(long? matricula = null, string? nome = null, 
         string? departamento = null, string? cargo = null, string? role = null);
-    Task<bool> ExistActiveUserMatricula(int matricula);
+    Task<bool> ExistActiveUserMatricula(long matricula);
     Task<Entities.Usuario?> GetUserByMatricula(long matricula);
     Task<Entities.Usuario?> GetById(long id);
