+++ src/Ativos.Infrastructure/DataAccess/Repositories/UsuariosRepository.cs
@@ -39,12 +39,12 @@ internal class UsuariosRepository : IUsuariosWriteOnlyRepository, IUsuariosReadO
         return await query.ToListAsync();
     }
 
-    public async Task<bool> ExistActiveUserMatricula(int matricula)
+    public async Task<bool> ExistActiveUserMatricula(long matricula)
     {
         return await _dbContext.Usuario.AnyAsync(user => user.Matricula.Equals(matricula)); //contem matricula
     }
 
-    public async Task<Usuario?> GetUserByMatricula(int matricula)
+    public async Task<Usuario?> GetUserByMatricula(long matricula)
     {
         return await _dbContext.Usuario.AsNoTracking().FirstOrDefaultAsync(usuario => usuario.Matricula.Equals(matricula));
     }
